refactor metadata and library handling
to make sure they are available at configuration time

+++ src/main/java/net/ornithemc/keratin/api/task/decompiling/DecompileTask.java
@@ -24,7 +24,9 @@ abstract class DecompileTask extends MinecraftTask {
 		OrnitheFilesAPI files = keratin.getFiles();
 
 		File src = files.getDecompiledSourceDirectory();
-		FileUtils.forceDelete(src);
+		if (src.exists()) {
+			FileUtils.forceDelete(src);
+		}
 
 		getProject().javaexec(javaexec -> {
 			javaexec.classpath(getProject().getConfigurations().getByName(Configurations.DECOMPILE_CLASSPATH));
